pyproject.toml
[project]
name = "worker-controller-demo"
version = "0.1.0"
description = "Temporal Worker Controller demo: versioned Python workers + dev UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "kubernetes>=31.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
demo-api = "api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["activity", "workflows", "worker", "api"]

[tool.uv]
package = true
